+++ filter.go
@@ -32,9 +32,19 @@ const (
 	// Unknown filters for an unknown (empty) value.
 	// This is a synonym for "null".
 	Unknown = "null"
+
+	// DirectionASC sorts results in ascending order.
+	DirectionASC = "ASC"
+
+	// DirectionDESC sorts results in descending order.
+	DirectionDESC = "DESC"
 )
 
+// Period is used to group results.
 type Period int
+
+// Direction is used to sort results.
+type Direction string
 
 // NullClient is a placeholder for no client (0).
 var NullClient = int64(0)
@@ -148,6 +158,16 @@ type Filter struct {
 	// EventMeta filters for event metadata.
 	EventMeta map[string]string
 
+	// Search searches the results for given fields and inputs.
+	Search []Search
+
+	// Sort sorts the results.
+	// This will overwrite the default order provided by the Analyzer.
+	Sort []Sort
+
+	// Offset limits the number of results. Less or equal to zero means no offset.
+	Offset int
+
 	// Limit limits the number of results. Less or equal to zero means no limit.
 	Limit int
 
@@ -164,6 +184,19 @@ type Filter struct {
 
 	eventFilter bool
 	minIsBot    uint8
+}
+
+// Search filters results by searching for the given input for given field.
+// The field needs to contain the search string and is performed case-insensitively.
+type Search struct {
+	Field Field
+	Input string
+}
+
+// Sort sorts results by a field and direction.
+type Sort struct {
+	Field     Field
+	Direction Direction
 }
 
 // NewFilter creates a new filter for given client ID.
@@ -209,21 +242,29 @@ func (filter *Filter) validate() {
 		filter.PathPattern = ""
 	}
 
+	for i := 0; i < len(filter.Search); i++ {
+		filter.Search[i].Input = strings.TrimSpace(filter.Search[i].Input)
+	}
+
+	if filter.Offset < 0 {
+		filter.Offset = 0
+	}
+
 	if filter.Limit < 0 {
 		filter.Limit = 0
 	}
 }
 
-func (filter *Filter) buildQuery(fields, groupBy, orderBy []field) ([]any, string) {
+func (filter *Filter) buildQuery(fields, groupBy, orderBy []Field) ([]any, string) {
 	table := filter.table()
 	args := make([]any, 0)
 	var query strings.Builder
 
-	if filter.Period != PeriodDay && filter.fieldsContain(fields, fieldDay.name) {
+	if filter.Period != PeriodDay && filter.fieldsContain(fields, FieldDay.name) {
 		query.WriteString(`SELECT `)
 
 		for i := range fields {
-			if fields[i] != fieldDay {
+			if fields[i] != FieldDay {
 				query.WriteString(fmt.Sprintf(`%s %s,`, fields[i].queryPeriod, fields[i].name))
 			}
 		}
@@ -238,7 +279,7 @@ func (filter *Filter) buildQuery(fields, groupBy, orderBy []field) ([]any, strin
 		}
 	}
 
-	if table == "event" || filter.Path != "" || filter.PathPattern != "" || filter.fieldsContain(fields, fieldPath.name) {
+	if table == "event" || filter.Path != "" || filter.PathPattern != "" || filter.fieldsContain(fields, FieldPath.name) {
 		if table == "session" {
 			table = "page_view"
 		}
@@ -248,10 +289,10 @@ func (filter *Filter) buildQuery(fields, groupBy, orderBy []field) ([]any, strin
 		if filter.minIsBot > 0 ||
 			filter.EntryPath != "" ||
 			filter.ExitPath != "" ||
-			filter.fieldsContain(fields, fieldBounces.name) ||
-			filter.fieldsContain(fields, fieldViews.name) ||
-			filter.fieldsContain(fields, fieldEntryPath.name) ||
-			filter.fieldsContain(fields, fieldExitPath.name) {
+			filter.fieldsContain(fields, FieldBounces.name) ||
+			filter.fieldsContain(fields, FieldViews.name) ||
+			filter.fieldsContain(fields, FieldEntryPath.name) ||
+			filter.fieldsContain(fields, FieldExitPath.name) {
 			filterArgs, filterQuery := filter.joinSessions(table, fields)
 			args = append(args, filterArgs...)
 			query.WriteString(filterQuery)
@@ -287,7 +328,7 @@ func (filter *Filter) buildQuery(fields, groupBy, orderBy []field) ([]any, strin
 		}
 	}
 
-	if filter.Period != PeriodDay && filter.fieldsContain(fields, fieldDay.name) {
+	if filter.Period != PeriodDay && filter.fieldsContain(fields, FieldDay.name) {
 		switch filter.Period {
 		case PeriodWeek:
 			query.WriteString(`) GROUP BY week ORDER BY week ASC`)
@@ -302,7 +343,7 @@ func (filter *Filter) buildQuery(fields, groupBy, orderBy []field) ([]any, strin
 	return args, query.String()
 }
 
-func (filter *Filter) joinPageViewFields(args *[]any, fields []field) string {
+func (filter *Filter) joinPageViewFields(args *[]any, fields []Field) string {
 	var out strings.Builder
 
 	for i := range fields {
@@ -324,7 +365,7 @@ func (filter *Filter) joinPageViewFields(args *[]any, fields []field) string {
 	return str[:len(str)-1]
 }
 
-func (filter *Filter) joinSessionFields(args *[]any, fields []field) string {
+func (filter *Filter) joinSessionFields(args *[]any, fields []Field) string {
 	var out strings.Builder
 
 	for i := range fields {
@@ -343,7 +384,7 @@ func (filter *Filter) joinSessionFields(args *[]any, fields []field) string {
 	return str[:len(str)-1]
 }
 
-func (filter *Filter) joinSessions(table string, fields []field) ([]any, string) {
+func (filter *Filter) joinSessions(table string, fields []Field) ([]any, string) {
 	path, pathPattern, eventName, eventMetaKey, eventMeta := filter.Path, filter.PathPattern, filter.EventName, filter.EventMetaKey, filter.EventMeta
 	filter.Path, filter.PathPattern, filter.EventName, filter.EventMetaKey, filter.EventMeta = "", "", "", "", nil
 	filterArgs, filterQuery := filter.query(true)
@@ -351,31 +392,31 @@ func (filter *Filter) joinSessions(table string, fields []field) ([]any, string)
 	sessionFields := make([]string, 0, 6)
 	groupBy := make([]string, 0, 4)
 
-	if filter.fieldsContain(fields, fieldEntryPath.name) {
-		sessionFields = append(sessionFields, fieldEntryPath.name)
-		groupBy = append(groupBy, fieldEntryPath.name)
+	if filter.fieldsContain(fields, FieldEntryPath.name) {
+		sessionFields = append(sessionFields, FieldEntryPath.name)
+		groupBy = append(groupBy, FieldEntryPath.name)
 	}
 
-	if filter.fieldsContain(fields, fieldExitPath.name) {
-		sessionFields = append(sessionFields, fieldExitPath.name)
-		groupBy = append(groupBy, fieldExitPath.name)
+	if filter.fieldsContain(fields, FieldExitPath.name) {
+		sessionFields = append(sessionFields, FieldExitPath.name)
+		groupBy = append(groupBy, FieldExitPath.name)
 	}
 
-	if filter.fieldsContainByQuerySession(fields, fieldEntryTitle.querySessions) {
-		sessionFields = append(sessionFields, fieldEntryTitle.querySessions)
-		groupBy = append(groupBy, fieldEntryTitle.querySessions)
+	if filter.fieldsContainByQuerySession(fields, FieldEntryTitle.querySessions) {
+		sessionFields = append(sessionFields, FieldEntryTitle.querySessions)
+		groupBy = append(groupBy, FieldEntryTitle.querySessions)
 	}
 
-	if filter.fieldsContainByQuerySession(fields, fieldExitTitle.querySessions) {
-		sessionFields = append(sessionFields, fieldExitTitle.querySessions)
-		groupBy = append(groupBy, fieldExitTitle.querySessions)
+	if filter.fieldsContainByQuerySession(fields, FieldExitTitle.querySessions) {
+		sessionFields = append(sessionFields, FieldExitTitle.querySessions)
+		groupBy = append(groupBy, FieldExitTitle.querySessions)
 	}
 
-	if filter.fieldsContain(fields, fieldBounces.name) {
+	if filter.fieldsContain(fields, FieldBounces.name) {
 		sessionFields = append(sessionFields, "sum(is_bounce*sign) is_bounce")
 	}
 
-	if filter.fieldsContain(fields, fieldViews.name) {
+	if filter.fieldsContain(fields, FieldViews.name) {
 		sessionFields = append(sessionFields, "sum(page_views*sign) page_views")
 	}
 
@@ -412,7 +453,7 @@ func (filter *Filter) joinSessions(table string, fields []field) ([]any, string)
 	return filterArgs, query
 }
 
-func (filter *Filter) joinGroupBy(fields []field) string {
+func (filter *Filter) joinGroupBy(fields []Field) string {
 	var out strings.Builder
 
 	for i := range fields {
@@ -423,7 +464,16 @@ func (filter *Filter) joinGroupBy(fields []field) string {
 	return str[:len(str)-1]
 }
 
-func (filter *Filter) joinOrderBy(args *[]any, fields []field) string {
+func (filter *Filter) joinOrderBy(args *[]any, fields []Field) string {
+	if len(filter.Sort) > 0 {
+		fields = make([]Field, 0, len(filter.Sort))
+
+		for i := range filter.Sort {
+			filter.Sort[i].Field.queryDirection = string(filter.Sort[i].Direction)
+			fields = append(fields, filter.Sort[i].Field)
+		}
+	}
+
 	var out strings.Builder
 
 	for i := range fields {
@@ -491,7 +541,7 @@ func (filter *Filter) queryTime(filterBots bool) ([]any, string) {
 }
 
 func (filter *Filter) queryFields() ([]any, string) {
-	n := 25 + len(filter.EventMeta) // maximum number of fields + one for bot filter + meta data fields
+	n := 25 + len(filter.EventMeta) + len(filter.Search) // maximum number of fields + one for bot filter + metadata fields + search fields
 	args := make([]any, 0, n)
 	queryFields := make([]string, 0, n)
 	filter.appendQuery(&queryFields, &args, "path", filter.Path)
@@ -519,6 +569,11 @@ func (filter *Filter) queryFields() ([]any, string) {
 	filter.queryPlatform(&queryFields)
 	filter.queryPathPattern(&queryFields, &args)
 	filter.appendQueryMeta(&queryFields, &args, filter.EventMeta)
+
+	for i := range filter.Search {
+		filter.appendQuerySearch(&queryFields, &args, filter.Search[i].Field.name, filter.Search[i].Input)
+	}
+
 	return args, strings.Join(queryFields, "AND ")
 }
 
@@ -618,7 +673,7 @@ func (filter *Filter) appendField(fields *[]string, field, value string) {
 	}
 }
 
-func (filter *Filter) fieldsContain(haystack []field, needle string) bool {
+func (filter *Filter) fieldsContain(haystack []Field, needle string) bool {
 	for i := range haystack {
 		if haystack[i].name == needle {
 			return true
@@ -628,7 +683,7 @@ func (filter *Filter) fieldsContain(haystack []field, needle string) bool {
 	return false
 }
 
-func (filter *Filter) fieldsContainByQuerySession(haystack []field, needle string) bool {
+func (filter *Filter) fieldsContainByQuerySession(haystack []Field, needle string) bool {
 	for i := range haystack {
 		if haystack[i].querySessions == needle {
 			return true
@@ -649,7 +704,9 @@ func (filter *Filter) withFill() ([]any, string) {
 }
 
 func (filter *Filter) withLimit() string {
-	if filter.Limit > 0 {
+	if filter.Limit > 0 && filter.Offset > 0 {
+		return fmt.Sprintf("LIMIT %d OFFSET %d ", filter.Limit, filter.Offset)
+	} else if filter.Limit > 0 {
 		return fmt.Sprintf("LIMIT %d ", filter.Limit)
 	}
 
@@ -691,6 +748,20 @@ func (filter *Filter) appendQuery(queryFields *[]string, args *[]any, field, val
 		}
 
 		*args = append(*args, filter.nullValue(value))
+		*queryFields = append(*queryFields, fmt.Sprintf(comparator, field))
+	}
+}
+
+func (filter *Filter) appendQuerySearch(queryFields *[]string, args *[]any, field, value string) {
+	if value != "" {
+		comparator := "ilike(%s, ?) = TRUE "
+
+		if strings.HasPrefix(value, "!") {
+			value = value[1:]
+			comparator = "ilike(%s, ?) = FALSE "
+		}
+
+		*args = append(*args, fmt.Sprintf("%%%s%%", value))
 		*queryFields = append(*queryFields, fmt.Sprintf(comparator, field))
 	}
 }